api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
elasticsearch[async]==8.17.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
modal==0.63.93
